Avoid warning for unused `test::Bencher`
When running `cargo build` rustc produces the warning:

    src/main.rs:113:5: 113:18 warning: unused import, #[warn(unused_imports)] on by default
    src/main.rs:113 use test::Bencher;

This is because functions marked with `#[bench]` are not compiled
during a non-test / non-bench build, leaving `test::Bencher` unused in
such a build.

In this commit we conditionally compile the include of `test::Bencher`
only when performing a test build (which are used for benchmarking).